requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-tinymce==4.1.0
django-unfold==0.40.0
psycopg2-binary>=2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
requests>=2.32.0
sqlparse==0.5.1
asgiref==3.8.1
